# Reduce locks during decision logging