### Why the changes in this PR are needed?

Decision logging is heavy on mutexes which increases the latency of decisions (see #5724) and also limits the parallelism. This PR removes two of the three Mutexes and replaces them with Golang structures for those use cases

### What are the changes in this PR?

Mutexes are replaced with `sync.Once` which has an optimised call path if the prepared queries are already set and only locks on the first occurrence.