@philipaconrad 2024-06-10
Okay! So I found an existing masking benchmark that seems to be illustrative of the improvements here: `BenchmarkMaskingErase`. It's dramatically better for illustrating the PR's effects than doing a mutex profile against the OPA server as a whole-- I found that at least for console logging the Logrus mutexes drown out any effect from this change! :sweat:

Here's how I ran my benchmarks:
```
git fetch origin pull/6797/head:remove_locks_decision_logs
git checkout main
go test -mutexprofile mtx.prof -run=^$ -tags opa_wasm,slow,e2e,noisy,use_opa_fork,no_license -bench ^BenchmarkMaskingErase$ github.com/open-policy-agent/opa/plugins/logs -count=30 > main.txt
git checkout remove_locks_decision_logs
go test -mutexprofile mtx.prof -run=^$ -tags opa_wasm,slow,e2e,noisy,use_opa_fork,no_license -bench ^BenchmarkMaskingErase$ github.com/open-policy-agent/opa/plugins/logs -count=30 > pr.txt
benchstat main.txt pr.txt
```
And here's the benchstat results from my `Core i7-8650U` Thinkpad:
```
name            old time/op  new time/op  delta
MaskingErase-8  16.4µs ± 5%  15.9µs ± 4%  -3.38%  (p=0.000 n=29+28)
```

So, I think that's a small, but solid bit of evidence in favor of this PR improving things! :smile: