@mjungsbluth 2024-06-12
One small observation: Before the change an error would set the query to `rego.PreparedEvalQuery{}`, now it is nil. I think this is generally nicer and idiomatic Go but it is a change. Was this intentional?

And thanks for ironing out the kinks :)