@johanfylling 2024-06-12
@mjungsbluth, I think we can chalk that one up as unconsciously intentional 😄. Since there is an initial state of `prepareOnce` where neither `preparedQuery` or `err` has been assigned values, it makes sense to internally use field types that can be `nil` to reflect being unassigned.